# packages/api-contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/registrations": {
      "post": {
        "operationId": "IdentityController_register_api/v1",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationResourceDto"
                }
              }
            }
          }
        },
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/sessions": {
      "post": {
        "operationId": "IdentityController_login_api/v1",
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResourceDto"
                }
              }
            }
          }
        },
        "tags": [
          "identity"
        ]
      }
    },
...
    "/api/v1/sessions/current": {
      "delete": {
        "operationId": "IdentityController_logoutCurrent_api/v1",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/users/me": {
      "get": {
        "operationId": "IdentityController_current_api/v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserResourceDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/users/me/ai-provider-consent": {
      "get": {
        "operationId": "IdentityController_getProviderConsent_api/v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiProviderConsentResourceDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "identity"
        ]
      },
      "post": {
        "operationId": "IdentityController_acceptProviderConsent_api/v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptAiProviderConsentRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiProviderConsentResourceDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/email-verification-requests": {
      "post": {
        "operationId": "IdentityController_requestVerification_api/v1",
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityAcceptedResourceDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/email-verifications": {
      "post": {
        "operationId": "IdentityController_confirmEmail_api/v1",
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailVerificationResourceDto"
                }
              }
            }
          }
        },
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/password-reset-requests": {
      "post": {
        "operationId": "IdentityController_requestReset_api/v1",
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityAcceptedResourceDto"
                }
              }
            }
          }
        },
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/password-resets": {
      "post": {
        "operationId": "IdentityController_confirmReset_api/v1",
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordResetResourceDto"
                }
              }
            }
          }
        },
        "tags": [
          "identity"
        ]
      }
    },
...
  },
  "info": {
    "title": "ATLAS API",
    "description": "",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "ConsentAcceptanceDto": {
        "type": "object",
        "properties": {
          "consentType": {
            "type": "string",
            "enum": [
              "terms",
              "privacy"
            ]
          },
          "documentVersion": {
            "type": "string",
            "example": "v1"
          },
          "accepted": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "consentType",
          "documentVersion",
          "accepted"
        ]
      },
      "RegistrationRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "person@example.com",
            "maxLength": 254
          },
          "password": {
            "type": "string",
            "minLength": 12,
            "maxLength": 128,
            "writeOnly": true
          },
          "ageConfirmed": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "consents": {
            "minItems": 2,
            "maxItems": 2,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConsentAcceptanceDto"
            }
          }
        },
        "required": [
          "email",
          "password",
          "ageConfirmed",
          "consents"
        ]
      },
      "RegistrationResourceDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "onboardingStatus": {
            "type": "string",
            "enum": [
              "registered",
              "profileReady",
              "personaReady",
              "completed"
            ]
          },
          "sessionExpiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "csrfToken": {
            "type": "string"
          },
          "emailVerified": {
            "type": "boolean"
          }
        },
        "required": [
          "userId",
          "onboardingStatus",
          "sessionExpiresAt",
          "csrfToken",
          "emailVerified"
        ]
      },
      "CreateSessionRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "person@example.com",
            "maxLength": 254
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "writeOnly": true
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "SessionResourceDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "onboardingStatus": {
            "type": "string",
            "enum": [
              "registered",
              "profileReady",
              "personaReady",
              "completed"
            ]
          },
          "csrfToken": {
            "type": "string"
          },
          "emailVerified": {
            "type": "boolean"
          }
        },
        "required": [
          "userId",
          "expiresAt",
          "onboardingStatus",
          "csrfToken",
          "emailVerified"
        ]
      },
      "CurrentUserResourceDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "onboardingStatus": {
            "type": "string",
            "enum": [
              "registered",
              "profileReady",
              "personaReady",
              "completed"
            ]
          },
          "emailVerified": {
            "type": "boolean"
          }
        },
        "required": [
          "userId",
          "onboardingStatus",
          "emailVerified"
        ]
      },
      "AiProviderConsentResourceDto": {
        "type": "object",
        "properties": {
          "accepted": {
            "type": "boolean"
          },
          "currentVersion": {
            "type": "string"
          },
          "acceptedVersion": {
            "type": "string",
            "nullable": true
          },
          "disclosure": {
            "type": "string"
          }
        },
        "required": [
          "accepted",
          "currentVersion",
          "acceptedVersion",
          "disclosure"
        ]
      },
      "AcceptAiProviderConsentRequestDto": {
        "type": "object",
        "properties": {
          "documentVersion": {
            "type": "string"
          },
          "accepted": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "documentVersion",
          "accepted"
        ]
      },
      "IdentityAcceptedResourceDto": {
        "type": "object",
        "properties": {
          "accepted": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "accepted"
        ]
      },
      "VerifyEmailRequestDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "minLength": 20,
            "writeOnly": true
          }
        },
        "required": [
          "token"
        ]
      },
      "EmailVerificationResourceDto": {
        "type": "object",
        "properties": {
          "emailVerified": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "emailVerified"
        ]
      },
      "PasswordResetRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "person@example.com",
            "maxLength": 254
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordRequestDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "minLength": 20,
            "writeOnly": true
          },
          "newPassword": {
            "type": "string",
            "minLength": 12,
            "maxLength": 128,
            "writeOnly": true
          }
        },
        "required": [
          "token",
          "newPassword"
        ]
      },
      "PasswordResetResourceDto": {
        "type": "object",
        "properties": {
          "passwordReset": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "passwordReset"
        ]
      },
...
    }
  }
}